+++ Core_Java/mockito/meus-numeros-api/src/test/java/br/dev/MockUsingAt.java
@@ -1,0 +1,32 @@
+package br.dev;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.MockitoAnnotations;
+import org.mockito.Spy;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.mockito.Mockito.when;
+
+class MockUsingAt {
+
+    @Mock
+    List<String> list;
+
+    @BeforeEach
+    void setUp() {
+        MockitoAnnotations.openMocks(this);
+    }
+
+    @Test
+    void testMockList() {
+        Mockito.when(list.get(0)).thenReturn("mocked");
+        assertEquals("mocked", list.get(0));
+    }
+}
